Move resolution utils into their own file
